+++ src/main/java/ru/skishop/service/OrderService.java
@@ -7,7 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.skishop.dto.OrderDto;
 import ru.skishop.dto.OrderItemDto;
 import ru.skishop.entity.*;
-import ru.skishop.exceptionHandler.EntityExistException;
 import ru.skishop.exceptionHandler.NotFoundException;
 import ru.skishop.mapper.OrderItemMapper;
 import ru.skishop.mapper.OrderMapper;
@@ -36,7 +35,7 @@ public class OrderService {
         return orderList.stream().map(orderMapper::toOrderDto).collect(Collectors.toList());
     }
 
-    public OrderDto findOrderById(Long orderId){
+    public OrderDto findOrderById(Long orderId) {
         Order order = orderRepository.findOrderById(orderId);
         return orderMapper.toOrderDto(order);
     }
@@ -45,13 +44,17 @@ public class OrderService {
     public OrderDto create() {
         Long userId = currentUser.getId();
 
+        List<UserBasketItem> userBasketItemDtoList = userBasketItemService.getBasketForCurrentUser();
+        if (userBasketItemDtoList.isEmpty()) {
+            log.info("OrderService: Basket is empty");
+            throw new NotFoundException("Basket is empty for userId = " + userId);
+        }
+        List<OrderItem> orderItemList = userBasketItemDtoList.stream().map(orderItemMapper::toOrderItem).collect(Collectors.toList());
+
         Order order = new Order();
         order.setDate(Instant.now());
         order.setUser(new User(userId));
         order = orderRepository.save(order);
-
-        List<UserBasketItem> userBasketItemDtoList = userBasketItemService.getBasketForCurrentUser();
-        List<OrderItem> orderItemList = userBasketItemDtoList.stream().map(orderItemMapper::toOrderItem).collect(Collectors.toList());
 
         Order finalOrder = order;
         orderItemList.forEach(orderItem -> orderItem.setOrder(new Order(finalOrder.getId())));
@@ -68,23 +71,22 @@ public class OrderService {
     }
 
     @Transactional
-    public OrderDto edit(Long orderId, Long skiId, int skiAmount){
+    public OrderDto edit(Long orderId, Long skiId, int skiAmount) {
         if (!orderRepository.existsById(orderId)) {
             log.info("OrderService: Not found order by id = {}", orderId);
             throw new NotFoundException("Not found order by id = " + orderId);
         }
-
         orderItemService.edit(orderId, skiId, skiAmount);
         Order order = orderRepository.getById(orderId);
         return orderMapper.toOrderDto(order);
     }
 
     @Transactional
-    public void delete(Long orderId){
-//        if (!orderRepository.existsById(orderId)) {
-//            log.info("OrderService: Not found order by id = {}", orderId);
-//            throw new NotFoundException("Not found order by id = " + orderId);
-//        }
+    public void delete(Long orderId) {
+        if (!orderRepository.existsById(orderId)) {
+            log.info("OrderService: Not found order by id = {}", orderId);
+            throw new NotFoundException("Not found order by id = " + orderId);
+        }
         orderItemService.delete(orderId);
         orderRepository.deleteById(orderId);
     }
